chart2svgを別ファイルに
